python/requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
streamlit>=1.20.0
Pillow>=9.0.0
scipy>=1.7.0
joblib>=1.1.0
